[NTFS] Implement GetVolumeBitmap().
This allows NFI to find which file is at which LCN: http://www.heisspiter.net/~Pierre/rostests/NTFS_NFI_LCN.png

So far, NFI is not as verbose as it is on Windows in such cases, because FILE_OPEN_BY_FILE_ID is *still* unimplemented (you should consider implementing it Pierre ;-)).
I checked with Windows, the correct file is found though!

CORE-8725


git-svn-id: svn://svn.reactos.org/reactos/trunk@65645 97493ccd-5924-5043-b1f5-66cb403b36ce